moved the anonymous class to a separate file. small style fix.
